3_dockerize/pulsar-testing/requirements.txt
pulsar-client==2.3.2

